+++ src/Pyro4/naming.py
@@ -10,231 +10,26 @@ import re
 import logging
 import socket
 import sys
-from Pyro4.threadutil import RLock, Lock, Thread
+from Pyro4.threadutil import RLock, Thread
 from Pyro4.errors import NamingError, PyroError, ProtocolError
 from Pyro4 import core, socketutil
 import Pyro4.constants
 
-__all__ = ["locateNS", "resolve", "startNS"]
+__all__ = ["locateNS", "resolve", "startNS", "startNSloop", "MemoryStorage"]
 
 if sys.version_info >= (3, 0):
     basestring = str
 
 log = logging.getLogger("Pyro4.naming")
 
-try:
-    import anydbm as dbm   # python 2
-except ImportError:
-    try:
-        import dbm   # python 3
-    except ImportError:
-        pass
-except Exception as x:
-    # pypy can generate a distutils error somehow if dbm is not available
-    pass
 
-
-class SqlStorage(object):
-    """
-    Sqlite-based storage.
-    """
-    def __init__(self, dbmfile):
-        raise NotImplementedError("Sqlite storage not yet implemented")   # XXX implement this :)
-
-
-class DbmStorage(object):
-    """
-    Storage implementation that uses a persistent dbm file.
-    Because dbm only supports strings as key/value, we encode/decode them in utf-8.
-    Dbm files cannot be accessed concurrently, so a strict concurrency model
-    is used where only one operation is processed at the same time
-    (this is very slow when compared to the in-memory storage)
-    """
-    def __init__(self, dbmfile):
-        self.dbmfile = dbmfile
-        db = dbm.open(self.dbmfile, "c", mode=0o600)
-        db.close()
-        self.lock = Lock()
-
-    def __getitem__(self, item):
-        item = item.encode("utf-8")
-        with self.lock:
-            try:
-                db = dbm.open(self.dbmfile)
-                try:
-                    return db[item].decode("utf-8")
-                finally:
-                    db.close()
-            except dbm.error as e:
-                raise NamingError("dbm error in getitem: "+str(e))
-
-    def __setitem__(self, key, value):
-        key = key.encode("utf-8")
-        value = value.encode("utf-8")
-        with self.lock:
-            try:
-                db = dbm.open(self.dbmfile, "w")
-                try:
-                    db[key] = value
-                finally:
-                    db.close()
-            except dbm.error as e:
-                raise NamingError("dbm error in setitem: "+str(e))
-
-    def __len__(self):
-        with self.lock:
-            try:
-                db = dbm.open(self.dbmfile)
-                try:
-                    return len(db)
-                finally:
-                    db.close()
-            except dbm.error as e:
-                raise NamingError("dbm error in len: "+str(e))
-
-    def __contains__(self, item):
-        item = item.encode("utf-8")
-        with self.lock:
-            try:
-                db = dbm.open(self.dbmfile)
-                try:
-                    return item in db
-                finally:
-                    db.close()
-            except dbm.error as e:
-                raise NamingError("dbm error in contains: "+str(e))
-
-    def __delitem__(self, key):
-        key = key.encode("utf-8")
-        with self.lock:
-            try:
-                db = dbm.open(self.dbmfile, "w")
-                try:
-                    del db[key]
-                finally:
-                    db.close()
-            except dbm.error as e:
-                raise NamingError("dbm error in delitem: "+str(e))
-
-    def __iter__(self):
-        with self.lock:
-            try:
-                db = dbm.open(self.dbmfile)
-                try:
-                    return iter([key.decode("utf-8") for key in db.keys()])
-                finally:
-                    db.close()
-            except dbm.error as e:
-                raise NamingError("dbm error in iter: "+str(e))
-
-    def clear(self):
-        with self.lock:
-            try:
-                db = dbm.open(self.dbmfile, "w")
-                try:
-                    if hasattr(db, "clear"):
-                        db.clear()
-                    else:
-                        for key in db.keys():
-                            del db[key]
-                finally:
-                    db.close()
-            except dbm.error as e:
-                raise NamingError("dbm error in clear: "+str(e))
-
-    def __getattr__(self, item):
-        raise NotImplementedError("DbmStorage doesn't implement method/attribute '"+item+"'")
-
-    def optimized_prefix_list(self, prefix):
-        with self.lock:
-            try:
-                db = dbm.open(self.dbmfile)
-                try:
-                    result = {}
-                    if hasattr(db, "items"):
-                        for key, value in db.items():
-                            key = key.decode("utf-8")
-                            if key.startswith(prefix):
-                                result[key] = value.decode("utf-8")
-                    else:
-                        for key in db.keys():
-                            keystr = key.decode("utf-8")
-                            if keystr.startswith(prefix):
-                                result[keystr] = db[key].decode("utf-8")
-                    return result
-                finally:
-                    db.close()
-            except dbm.error as e:
-                raise NamingError("dbm error in optimized_prefix_list: "+str(e))
-
-    def optimized_regex_list(self, regex):
-        try:
-            regex = re.compile(regex + "$")  # add end of string marker
-        except re.error:
-            x = sys.exc_info()[1]
-            raise NamingError("invalid regex: " + str(x))
-        with self.lock:
-            try:
-                db = dbm.open(self.dbmfile)
-                try:
-                    result = {}
-                    if hasattr(db, "items"):
-                        for key, value in db.items():
-                            key = key.decode("utf-8")
-                            if regex.match(key):
-                                result[key] = value.decode("utf-8")
-                    else:
-                        for key in db.keys():
-                            keystr = key.decode("utf-8")
-                            if regex.match(keystr):
-                                result[keystr] = db[key].decode("utf-8")
-                    return result
-                finally:
-                    db.close()
-            except dbm.error as e:
-                raise NamingError("dbm error in optimized_regex_list: "+str(e))
-
-    def remove_all(self, items):
-        with self.lock:
-            try:
-                db = dbm.open(self.dbmfile, "w")
-                try:
-                    for item in items:
-                        try:
-                            del db[item.encode("utf-8")]
-                        except KeyError:
-                            pass
-                finally:
-                    db.close()
-            except dbm.error as e:
-                raise NamingError("dbm error in remove_all: "+str(e))
-
-    def everything(self):
-        with self.lock:
-            try:
-                db = dbm.open(self.dbmfile)
-                try:
-                    result = {}
-                    if hasattr(db, "items"):
-                        for key, value in db.items():
-                            result[key.decode("utf-8")] = value.decode("utf-8")
-                    else:
-                        for key in db.keys():
-                            result[key.decode("utf-8")] = db[key].decode("utf-8")
-                    return result
-                finally:
-                    db.close()
-            except dbm.error as e:
-                raise NamingError("dbm error in everything: "+str(e))
-
-
-class DictStorage(dict):
+class MemoryStorage(dict):
     """
     Storage implementation that is just an in-memory dict.
     Stopping the nameserver will make the server instantly forget about everything.
     """
     def __init__(self, **kwargs):
-        super(DictStorage, self).__init__(**kwargs)
+        super(MemoryStorage, self).__init__(**kwargs)
 
     def optimized_prefix_list(self, prefix):
         return None
@@ -245,12 +40,15 @@ class DictStorage(dict):
     def everything(self):
         return self.copy()
 
-    def remove_all(self, items):
+    def remove_items(self, items):
         for item in items:
             try:
                 del self[item]
             except KeyError:
                 pass
+
+    def close(self):
+        pass
 
 
 class NameServer(object):
@@ -261,12 +59,9 @@ class NameServer(object):
     def __init__(self, storageProvider=None):
         self.storage = storageProvider
         if storageProvider is None:
-            self.storage = DictStorage()
+            self.storage = MemoryStorage()
             log.debug("using volatile in-memory dict storage")
         self.lock = RLock()
-
-    def clear(self):
-        self.storage.clear()      # very rarely needed! (mostly only in test cases)
 
     def count(self):
         return len(self.storage)
@@ -304,13 +99,13 @@ class NameServer(object):
             items = list(self.list(prefix=prefix).keys())
             if Pyro4.constants.NAMESERVER_NAME in items:
                 items.remove(Pyro4.constants.NAMESERVER_NAME)
-            self.storage.remove_all(items)
+            self.storage.remove_items(items)
             return len(items)
         if regex:
             items = list(self.list(regex=regex).keys())
             if Pyro4.constants.NAMESERVER_NAME in items:
                 items.remove(Pyro4.constants.NAMESERVER_NAME)
-            self.storage.remove_all(items)
+            self.storage.remove_items(items)
             return len(items)
         return 0
 
@@ -368,14 +163,16 @@ class NameServerDaemon(core.Daemon):
         storage = storage or "memory"
         if storage == "memory":
             log.debug("using volatile in-memory dict storage")
-            self.nameserver = NameServer(DictStorage())
+            self.nameserver = NameServer(MemoryStorage())
         elif storage.startswith("dbm:") and len(storage) > 4:
             dbmfile = storage[4:]
             log.debug("using persistent dbm storage in file %s", dbmfile)
+            from Pyro4.naming_storage import DbmStorage
             self.nameserver = NameServer(DbmStorage(dbmfile))
         elif storage.startswith("sql:") and len(storage) > 4:
             sqlfile = storage[4:]
             log.debug("using persistent sql storage in file %s", sqlfile)
+            from Pyro4.naming_storage import SqlStorage
             self.nameserver = NameServer(SqlStorage(sqlfile))
         else:
             raise ValueError("invalid storage type '%s'" % storage)
@@ -389,7 +186,9 @@ class NameServerDaemon(core.Daemon):
 
     def close(self):
         super(NameServerDaemon, self).close()
-        self.nameserver = None
+        if self.nameserver is not None:
+            self.nameserver.storage.close()
+            self.nameserver = None
 
     def __enter__(self):
         if not self.nameserver:
@@ -397,6 +196,8 @@ class NameServerDaemon(core.Daemon):
         return self
 
     def __exit__(self, exc_type, exc_value, traceback):
+        if self.nameserver is not None:
+            self.nameserver.storage.close()
         self.nameserver = None
         return super(NameServerDaemon, self).__exit__(exc_type, exc_value, traceback)
 
@@ -643,7 +444,7 @@ def main(args=None):
     parser.add_option("-n", "--host", dest="host", help="hostname to bind server on")
     parser.add_option("-p", "--port", dest="port", type="int", help="port to bind server on (0=random)")
     parser.add_option("-u", "--unixsocket", help="Unix domain socket name to bind server on")
-    parser.add_option("-s", "--storage", help="Storage system to use, default=memory", default="memory")
+    parser.add_option("-s", "--storage", help="Storage system to use (memory, dbm:file, sql:file)", default="memory")
     parser.add_option("", "--bchost", dest="bchost", help="hostname to bind broadcast server on (default is \"\")")
     parser.add_option("", "--bcport", dest="bcport", type="int",
                       help="port to bind broadcast server on (0=random)")